backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
python-dotenv==1.0.1